requirements.txt
boto3
fastapi
faker
uvicorn
Jinja2
requests
python-dotenv
botocore
ipython
pydantic
google-cloud-storage
azure-storage-blob
azure-keyvault-secrets
azure-identity
azure-mgmt-compute
google-compute-engine
google-cloud-dns
google-cloud-appengine-admin
mysql-connector-python
SQLAlchemy
pymongo